requirements.txt
langchain-core
langchain-community
langgraph
fastapi
uvicorn
langchain-groq
langchain-huggingface
numpy
pandas
twilio
python-dotenv
python-multipart
streamlit
transformers
sentence-transformers
faiss-cpu  
pyngrok